Dev hj (#114)
* mywishlist4

* mypageupdate5

* mypagelogin

---------

+++ 01_Backend/backend/src/main/java/org/example/backend/model/common/PdIdUserIdPk.java
@@ -5,17 +5,18 @@ import lombok.*;
 import java.io.Serializable;
 
 /**
- * packageName : org.example.backend.model.common
+ * packageName : org.example.backendproject.model.common
  * fileName : PdIdUserIdPk
- * author : kimtaewan
- * date : 2024-04-26
- * description : 상품 ID, 회원 ID 복합키
+ * author : SAMSUNG
+ * date : 2024-04-23
+ * description : 상품 ID, 회원 ID 복합키 클래스
  * 요약 :
  * <p>
  * ===========================================================
  * DATE            AUTHOR             NOTE
  * -----------------------------------------------------------
- * 2024-04-26         kimtaewan          최초 생성
+ * 2024-04-23         SAMSUNG          최초 생성
+
  */
 @Setter
 @Getter
@@ -24,6 +25,6 @@ import java.io.Serializable;
 @NoArgsConstructor
 @EqualsAndHashCode
 public class PdIdUserIdPk implements Serializable {
-    private Long pdId;
-    private String userId;
+    private Integer pdId;       // 상품 ID
+    private String userId;      // 회원 ID
 }
